# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Inter Finder API",
    "version": "1.0.0",
    "description": "API Documentation for the Inter Finder web application. This provides a clear contract for frontend and backend teams."
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/applications/interns": {
      "get": {
        "summary": "Get all applications for the logged-in intern",
        "tags": [
          "Applications"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Retrieves all applications submitted by the current intern",
        "responses": {
          "200": {
            "description": "Successfully retrieved applications"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (intern only)"
          }
        }
      }
    },
...
    "/auth/register/intern": {
      "post": {
        "summary": "Register a new user with the 'intern' role",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "password",
                  "phone",
                  "internshipType"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "John Intern"
                  },
                  "email": {
                    "type": "string",
                    "example": "john.intern@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  },
                  "phone": {
                    "type": "string",
                    "example": "123-456-7890"
                  },
                  "internshipType": {
                    "type": "string",
                    "enum": [
                      "free",
                      "paid"
                    ],
                    "example": "paid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Intern registered successfully. Returns a JWT token."
          },
          "400": {
            "description": "Bad Request (e.g., email already exists, missing fields)"
          }
        }
      }
    },
...
    "/auth/login/intern": {
      "post": {
        "summary": "Login for an existing intern",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "john.intern@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful. Returns a JWT token."
          },
          "401": {
            "description": "Unauthorized (Invalid credentials or user is not an intern)"
          }
        }
      }
    },
...
    "/interns/profile": {
      "get": {
        "summary": "Get current intern's profile",
        "tags": [
          "Interns"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved intern profile"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (intern only)"
          },
          "404": {
            "description": "Intern profile not found"
          }
        }
      },
      "put": {
        "summary": "Create or update intern profile",
        "tags": [
          "Interns"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "headline": {
                    "type": "string"
                  },
                  "bio": {
                    "type": "string"
                  },
                  "location": {
                    "type": "string"
                  },
                  "website": {
                    "type": "string"
                  },
                  "github": {
                    "type": "string"
                  },
                  "linkedin": {
                    "type": "string"
                  },
                  "twitter": {
                    "type": "string"
                  },
                  "education": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "school": {
                          "type": "string"
                        },
                        "degree": {
                          "type": "string"
                        },
                        "fieldOfStudy": {
                          "type": "string"
                        },
                        "from": {
                          "type": "string",
                          "format": "date"
                        },
                        "to": {
                          "type": "string",
                          "format": "date"
                        },
                        "current": {
                          "type": "boolean"
                        },
                        "description": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "skills": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "experience": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "title": {
                          "type": "string"
                        },
                        "company": {
                          "type": "string"
                        },
                        "location": {
                          "type": "string"
                        },
                        "from": {
                          "type": "string",
                          "format": "date"
                        },
                        "to": {
                          "type": "string",
                          "format": "date"
                        },
                        "current": {
                          "type": "boolean"
                        },
                        "description": {
                          "type": "string"
                        }
                      }
                    }
                  },
                  "resume": {
                    "type": "string"
                  },
                  "profilePicture": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Intern profile updated successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (intern only)"
          }
        }
      }
    },
    "/interns": {
      "get": {
        "summary": "Get all interns",
        "tags": [
          "Interns"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "select",
            "schema": {
              "type": "string"
            },
            "description": "Fields to select (comma-separated)"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            },
            "description": "Sort by fields (prefix with - for descending)"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Number of items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved interns",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "pagination": {
                      "type": "object"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Intern"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/interns/{id}": {
      "get": {
        "summary": "Get intern by ID",
        "tags": [
          "Interns"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Intern ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved intern",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Intern"
                }
              }
            }
          },
          "404": {
            "description": "Intern not found"
          }
        }
      }
    },
...
    "/interviews/intern": {
      "get": {
        "summary": "Get all interviews for the logged-in intern",
        "tags": [
          "Interviews"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved interviews"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not authorized (intern only)"
          }
        }
      }
    },
...
  },
...
}